+++ storage/url/url.go
@@ -40,12 +40,14 @@ const (
 // URL is the canonical representation of an object, either on local or remote
 // storage.
 type URL struct {
-	Type      urlType
-	Scheme    string
-	Bucket    string
-	Path      string
-	Delimiter string
-	Prefix    string
+	Type        urlType
+	Scheme      string
+	Bucket      string
+	Path        string
+	Delimiter   string
+	Prefix      string
+	VersionID   string
+	AllVersions bool
 
 	relativePath string
 	filter       string
@@ -58,6 +60,18 @@ type Option func(u *URL)
 func WithRaw(mode bool) Option {
 	return func(u *URL) {
 		u.raw = mode
+	}
+}
+
+func WithVersion(versionId string) Option {
+	return func(u *URL) {
+		u.VersionID = versionId
+	}
+}
+
+func WithAllVersions(isAllVersions bool) Option {
+	return func(u *URL) {
+		u.AllVersions = isAllVersions
 	}
 }
 
@@ -148,6 +162,11 @@ func (u *URL) IsPrefix() bool {
 // IsBucket returns true if the object url contains only bucket name
 func (u *URL) IsBucket() bool {
 	return u.IsRemote() && u.Path == ""
+}
+
+// IsVersioned returns true if the URL has versioning related values
+func (u *URL) IsVersioned() bool {
+	return u.AllVersions || u.VersionID != ""
 }
 
 // Absolute returns the absolute URL format of the object.
@@ -280,16 +299,19 @@ func (u *URL) setPrefixAndFilter() error {
 // Clone creates a copy of the receiver.
 func (u *URL) Clone() *URL {
 	return &URL{
-		Type:      u.Type,
-		Scheme:    u.Scheme,
-		Bucket:    u.Bucket,
-		Delimiter: u.Delimiter,
-		Path:      u.Path,
-		Prefix:    u.Prefix,
+		Type:        u.Type,
+		Scheme:      u.Scheme,
+		Bucket:      u.Bucket,
+		Path:        u.Path,
+		Delimiter:   u.Delimiter,
+		Prefix:      u.Prefix,
+		VersionID:   u.VersionID,
+		AllVersions: u.AllVersions,
 
 		relativePath: u.relativePath,
 		filter:       u.filter,
 		filterRegex:  u.filterRegex,
+		raw:          u.raw,
 	}
 }