+++ strutil/strutil.go
@@ -43,6 +43,17 @@ func JSON(v interface{}) string {
 	return string(bytes)
 }
 
+// CapitalizeFirstRune converts first rune to uppercase, and converts rest of
+// the string to lower case.
+func CapitalizeFirstRune(str string) string {
+	if str == "" {
+		return str
+	}
+	runes := []rune(str)
+	first, rest := runes[0], runes[1:]
+	return strings.ToUpper(string(first)) + strings.ToLower(string(rest))
+}
+
 // AddNewLineFlag adds a flag that allows . to match new line character "\n".
 // It assumes that the pattern does not have any flags.
 func AddNewLineFlag(pattern string) string {